Escape key correctly closes alert/confirm/dialog
An alert button only has an okay button, so in this case Esc hits OK.
Confirm and dialog have a cancel button, so it is more appropriate to
hit Cancel.

The listener for the Enter key has been removed since Enter should only
act on the element that has focus.